Made  the website more responsive
I added more relative units instead of absolute, so the responsiveness
is a lot better now. Main problem the page had is that it had fixed widths
for inputs, which made the page super unresponisive. I decided to go with
viewport widths on them, so they scale a lot better depending on the
viewport width.

Also made some barely noticable changes.